Solved MISRA violation 10.3

+++ FreeRTOS_Alarm/FreeRTOS_Alarm.ino
@@ -680,7 +680,7 @@ void taskMotionSensor(void *pvParameters)
         pin_pir=PIR1_PIN;
         virtual_pin=V1;
         str_code_blynk=str_code_blynk1;
-        position_pir=(uint32_t)POSITION_PIR1;
+        position_pir=(uint8_t)POSITION_PIR1;
     }
     else{
         pin_pir=PIR2_PIN;
